+++ Game/Client/DevilBerserkerSwapSkill.cpp
@@ -48,8 +48,8 @@ void DevilBerserkerSwapSkill::Update()
 			m_tUpTimer.Stop();
 			GET_SINGLE(Scenes)->GetActiveScene()->ShakeCameraAxis(0.1f, Vec3(0.f, 1000.f, 0.f));
 			Vec3 vMyPos = m_pSkul.lock()->GetPlayer().lock()->GetTransform()->GetPhysicalPosition();
-			GET_SINGLE(CollisionManager)->SetForceInLayer(LAYER_TYPE::PARTICLE, vMyPos, Vec3(300.f, 50.f, 0.f), Vec3(RANDOM(-300, 300), 300.f, 0.f));
-			GET_SINGLE(CollisionManager)->SetForceInMonsterAndTakeDamage(vMyPos, Vec3(300.f, 50.f, 0.f), Vec3(RANDOM(-300, 300), 300.f, 0.f), static_cast<float>(RANDOM(10, 30)), DAMAGE_TYPE::FROM_PLAYER_MELEE);
+			GET_SINGLE(CollisionManager)->SetForceInLayer(LAYER_TYPE::PARTICLE, vMyPos, Vec3(300.f, 50.f, 0.f), Vec3(static_cast<float>(RANDOM(-300, 300)), 300.f, 0.f));
+			GET_SINGLE(CollisionManager)->SetForceInMonsterAndTakeDamage(vMyPos, Vec3(300.f, 50.f, 0.f), Vec3(static_cast<float>(RANDOM(-300, 300)), 300.f, 0.f), static_cast<float>(RANDOM(10, 30)), DAMAGE_TYPE::FROM_PLAYER_MELEE);
 
 		}
 		else
